% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.vector.bdMatrix}
\alias{as.vector.bdMatrix}
\title{Coerce bdMatrix to vector}
\usage{
as.vector.bdMatrix(x, mode)
}
\arguments{
\item{x}{object of class bdMatrix}

\item{mode}{see \code{?as.vector}}
}
\description{
Coerce bdMatrix to vector
}

