% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_shared.R
\docType{methods}
\name{crossprod,bdMatrix,kroneckersumBlockMatrix-method}
\alias{crossprod,bdMatrix,kroneckersumBlockMatrix-method}
\title{crossprod function for crossproduct of a bdMatrix and a kroneckersumBlockMatrix}
\usage{
\S4method{crossprod}{bdMatrix,kroneckersumBlockMatrix}(x, y)
}
\arguments{
\item{x}{object of class \code{bdMatrix}}

\item{y}{object of class \code{kroneckersumBlockMatrix}}
}
\description{
crossprod function for crossproduct of a bdMatrix and a kroneckersumBlockMatrix
}
\details{
if \code{x} and \code{y} have the same block sizes, the crossprod function
calculates the crossproduct of each block separately
}

