% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_rowtensorBlockMatrix.R
\docType{methods}
\name{crossprod,rowtensorBlockMatrix,ANY-method}
\alias{crossprod,rowtensorBlockMatrix,ANY-method}
\title{crossprod method for rowtensorBlockMatrix object}
\usage{
\S4method{crossprod}{rowtensorBlockMatrix,ANY}(x)
}
\arguments{
\item{x}{object of class \code{rowtensorBlockMatrix}}
}
\description{
crossprod method for rowtensorBlockMatrix object
}

