% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_rowtensorBlockMatrix.R
\docType{methods}
\name{crossprod,rowtensorBlockMatrix,rowtensorBlockMatrix-method}
\alias{crossprod,rowtensorBlockMatrix,rowtensorBlockMatrix-method}
\title{crossprod method for rowtensorBlockMatrix objects}
\usage{
\S4method{crossprod}{rowtensorBlockMatrix,rowtensorBlockMatrix}(x, y)
}
\arguments{
\item{x}{object of class \code{rowtensorBlockMatrix}}

\item{y}{see \code{x}}
}
\description{
crossprod method for rowtensorBlockMatrix objects
}

