% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_shared.R
\docType{methods}
\name{\%*\%,rowtensorBlockMatrix,bdMatrix-method}
\alias{\%*\%,rowtensorBlockMatrix,bdMatrix-method}
\title{Matrix multiplication function for a rowtensorBlockMatrix and a bdMatrix}
\usage{
\S4method{\%*\%}{rowtensorBlockMatrix,bdMatrix}(x, y)
}
\arguments{
\item{x}{object of class \code{bdMatrix}}

\item{y}{object of class \code{kroneckersumBlockMatrix}}
}
\description{
Matrix multiplication function for a rowtensorBlockMatrix and a bdMatrix
}
\details{
if \code{x} and \code{y} have the same block sizes, the function
calculates the ordinary matrix product of each block separately
}

