% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods_kroneckersumBlockMatrix.R
\docType{class}
\name{kroneckersumBlockMatrix-class}
\alias{*,kroneckersumBlockMatrix,numeric-method}
\alias{*,numeric,kroneckersumBlockMatrix-method}
\alias{kroneckersumBlockMatrix-class}
\alias{nrow,kroneckersumBlockMatrix-method}
\title{Matrix class and methods for block structured kronecker sums}
\usage{
\S4method{*}{kroneckersumBlockMatrix,numeric}(e1, e2)

\S4method{*}{numeric,kroneckersumBlockMatrix}(e1, e2)

\S4method{nrow}{kroneckersumBlockMatrix}(x)
}
\arguments{
\item{e1}{see \code{x}}

\item{e2}{see \code{x}}

\item{x}{object of class \code{kroneckersumBlockMatrix}}
}
\description{
Matrix class and methods for block structured kronecker sums
}

