% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{kroneckersumBlockMatrix}
\alias{kroneckersumBlockMatrix}
\title{Constructor for kroneckersumBlockMatrix objects}
\usage{
kroneckersumBlockMatrix(X1, X2)
}
\arguments{
\item{X1}{a diagonal matrix of class \code{ddiMatrix}}

\item{X2}{a \code{Matrix} object}
}
\description{
Constructor for kroneckersumBlockMatrix objects
}
\examples{
matLeft = Diagonal(5)
matRight = Matrix(1:9, ncol=3)
kroneckersumBlockMatrix(matLeft, matRight)

}

