% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_shared.R
\docType{methods}
\name{+,bdMatrix,kroneckersumBlockMatrix-method}
\alias{+,bdMatrix,kroneckersumBlockMatrix-method}
\title{Function to sum a bdMatrix and a kroneckersumBlockMatrix}
\usage{
\S4method{+}{bdMatrix,kroneckersumBlockMatrix}(e1, e2)
}
\arguments{
\item{e1}{object of class \code{bdMatrix}}

\item{e2}{object of class \code{kroneckersumBlockMatrix}}
}
\description{
Function to sum a bdMatrix and a kroneckersumBlockMatrix
}
\details{
\code{x} and \code{y} must have the same number of blocks
WARNING: Sizes of blocks are not checked. This can lead to unwanted results.
}

