% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rowtensorBlockMatrix.R
\docType{methods}
\name{rankMatrix,rowtensorBlockMatrix-method}
\alias{rankMatrix,rowtensorBlockMatrix-method}
\title{rankMatrix extension for rowtensorBlockMatrix objects}
\usage{
\S4method{rankMatrix}{rowtensorBlockMatrix}(x, method, warn.t)
}
\arguments{
\item{x}{object of class rowtensorBlockMatrix or a numeric matrix}

\item{...}{further arguments passed to \code{Matrix::rankMatrix}}
}
\description{
rankMatrix extension for rowtensorBlockMatrix objects
}
\details{
if x is a numeric matrix \code{Matrix::rankMatrix} is called on x. Else the rank is computed
as product of number of colums of the \code{matLeft}-slot and the rank of the \code{matRight}-slot,
which is also calculated via Matrix::rankMatrix
}

