% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rowtensorBlockMatrix.R
\name{rankMatrix}
\alias{rankMatrix}
\title{rankMatrix extension for rowtensorBlockMatrix objects}
\usage{
rankMatrix(x, ...)
}
\arguments{
\item{x}{object of class rowtensorBlockMatrix or a numeric matrix}

\item{...}{further arguments passed to \code{rankMatrixMatrix}, which is the original \code{Matrix::rankMatrix} function}
}
\description{
rankMatrix extension for rowtensorBlockMatrix objects
}
\details{
if x is a numeric matrix the original \code{Matrix::rankMatrix} is called on x. 
Else the rank is computed #' as product of number of colums of the \code{matLeft}-slot and 
the rank of the \code{matRight}-slot, which is also calculated via \code{Matrix::rankMatrix}
}

