% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rowtensorBlockMatrix.R
\name{rankMatrix.rowtensorBlockMatrix}
\alias{rankMatrix.rowtensorBlockMatrix}
\title{rankMatrix function for rowtensorBlockMatrix}
\usage{
rankMatrix.rowtensorBlockMatrix(x, method = c("tolNorm2", "qr.R", "qrLINPACK",
  "qr", "useGrad", "maybeGrad"), warn.t = TRUE)
}
\description{
rankMatrix function for rowtensorBlockMatrix
}

