% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{rowtensorBlockMatrix-class}
\alias{rowtensorBlockMatrix-class}
\title{Matrix class for block structured row-wise tensor products}
\description{
Matrix class for block structured row-wise tensor products
}

