% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\name{rowtensorBlockMatrix}
\alias{rowtensorBlockMatrix}
\title{Constructor for rowtensorBlockMatrix objects}
\usage{
rowtensorBlockMatrix(X1, X2)
}
\arguments{
\item{X1}{a \code{bdMatrix} object}

\item{X2}{a \code{Matrix} object}
}
\description{
Constructor for rowtensorBlockMatrix objects
}

