% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bdMatrix.R
\docType{methods}
\name{solve,bdMatrix,bdMatrix-method}
\alias{solve,bdMatrix,bdMatrix-method}
\title{solve method for class bdMatrix}
\usage{
\S4method{solve}{bdMatrix,bdMatrix}(a, b, ...)
}
\arguments{
\item{a}{object of class \code{bdMatrix}}

\item{b}{object of class \code{bdMatrix}}

\item{...}{further arguments passed to \code{solve} (see \code{?solve})}
}
\description{
solve method for class bdMatrix
}
\details{
Due to the block diagonal structure of \code{a}, solving can be performed separately on block levels
if \code{a} only consists of quadratic blocks. However, \code{solve} does not exploit the second argument \code{b}
as it is done in solve with vectors or matrices. Instead, \code{solve(a,b)} simply calculates \code{solve(a)\%*\%b}
blockwise for corresponding blocks.
}

