% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bdMatrix.R
\docType{methods}
\name{svd,bdMatrix-method}
\alias{svd,bdMatrix-method}
\title{svd method for bdMatrix class}
\usage{
\S4method{svd}{bdMatrix}(x, nu = min(nrow(x), p = ncol(x)),
  nv = min(nrow(x), p = ncol(x)))
}
\arguments{
\item{x}{a object of class \code{bdMatrix} whose SVD decomposition is to be computed}

\item{nu}{the number of left singular vectors to be computed. This must between \code{0} and \code{n = nrow(x)}.}

\item{nv}{the number of right singular vectors to be computed. This must be between \code{0} and \code{p = ncol(x)}.}
}
\description{
svd method for bdMatrix class
}

