% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_bdMatrix.R
\docType{methods}
\name{t,bdMatrix-method}
\alias{t,bdMatrix-method}
\title{transpose method for bdMatrix objects}
\usage{
\S4method{t}{bdMatrix}(x)
}
\arguments{
\item{x}{object of class \code{bdMatrix}}
}
\description{
transpose method for bdMatrix objects
}

