% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{bdMatrix}
\alias{bdMatrix}
\title{Constructor for block diagonal matrices}
\usage{
bdMatrix(listOfBlocks)
}
\arguments{
\item{listOfBlocks}{a list of matrices or numeric vectors, which represent the blocks in the given order}
}
\description{
Constructor for block diagonal matrices
}

