% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{kroneckersumBlockMatrix-class}
\alias{kroneckersumBlockMatrix-class}
\title{Matrix class for block structured kronecker sums}
\description{
Matrix class for block structured kronecker sums
}

