% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rowtensorBlockMatrix.R
\name{rankMatrix.rowtensorBlockMatrix}
\alias{rankMatrix.rowtensorBlockMatrix}
\title{rankMatrix function for rowtensorBlockMatrix class}
\usage{
\method{rankMatrix}{rowtensorBlockMatrix}(x, ...)
}
\description{
rankMatrix function for rowtensorBlockMatrix class
}

