% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods_rowtensorBlockMatrix.R
\docType{class}
\name{rowtensorBlockMatrix-class}
\alias{*,rowtensorBlockMatrix,numeric-method}
\alias{[,rowtensorBlockMatrix,ANY,ANY,ANY-method}
\alias{\%*\%,rowtensorBlockMatrix,numeric-method}
\alias{abs,rowtensorBlockMatrix-method}
\alias{crossprod,rowtensorBlockMatrix,numeric-method}
\alias{dim,rowtensorBlockMatrix-method}
\alias{max,rowtensorBlockMatrix-method}
\alias{rowtensorBlockMatrix-class}
\title{Matrix class and methods for block structured row-wise tensor products}
\usage{
\S4method{[}{rowtensorBlockMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{crossprod}{rowtensorBlockMatrix,numeric}(x, y)

\S4method{\%*\%}{rowtensorBlockMatrix,numeric}(x, y)

\S4method{*}{rowtensorBlockMatrix,numeric}(e1, e2)

\S4method{dim}{rowtensorBlockMatrix}(x)

\S4method{abs}{rowtensorBlockMatrix}(x)

\S4method{max}{rowtensorBlockMatrix}(x)
}
\arguments{
\item{x}{object of class \code{rowtensorBlockMatrix}}

\item{i}{integer for subsetting rows of \code{x}}

\item{j}{integer for subsetting columns of \code{x}}

\item{...}{further arguments passed to ordinary subsetting functions}

\item{drop}{logical. If TRUE the result is coerced to the lowest possible dimension.}

\item{y}{see \code{x}}

\item{e1}{see \code{x}}

\item{e2}{see \code{x}}
}
\description{
Matrix class and methods for block structured row-wise tensor products
}

