% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rowtensorBlockMatrix}
\alias{rowtensorBlockMatrix}
\title{Constructor for rowtensorBlockMatrix objects}
\usage{
rowtensorBlockMatrix(X1, X2)
}
\arguments{
\item{X1}{a \code{bdMatrix} object}

\item{X2}{a \code{Matrix} object}
}
\description{
Constructor for rowtensorBlockMatrix objects
}
\examples{
matLeft = bdMatrix(list(c(1), matrix(c(1,1,1), ncol=1)))
matRight = Matrix(1:8, ncol=2)
rowtensorBlockMatrix(matLeft, matRight)

}

